--- src/main/java/org/ayty/model/Login.java
@@ -1,11 +1,0 @@
-package org.ayty.model;
-
-import lombok.Data;
-
-@Data
-public class Login {
-
-	private String username;
-	private String password;
-
-}
